tidy up path generation on path annotator to use data model
