+++ examples/library/surface_annotator.py
@@ -1,29 +1,16 @@
 import napari
-import skimage
+from skimage import data
 
-from napari_threedee.annotators._qt import QtSurfaceAnnotatorWidget
+import napari_threedee as n3d
 
-
+# create napari viewer
 viewer = napari.Viewer(ndisplay=3)
-blobs = skimage.data.binary_blobs(
-    length=64,
-    volume_fraction=0.1,
-    n_dim=3
-).astype(float)
 
-plane_parameters_z = {
-    'position': (32, 32, 32),
-    'normal': (1, 0, 0),
-    'thickness': 10,
-}
+# generate 3d image data
+blobs = data.binary_blobs(length=64, volume_fraction=0.1, n_dim=3).astype(float)
 
-plane_parameters_y = {
-    'position': (32, 32, 32),
-    'normal': (0, 1, 0),
-    'thickness': 10,
-}
-
-viewer.add_image(
+# add image layer to viewer (rendering as plane)
+image_layer = viewer.add_image(
     blobs,
     name='orange plane',
     rendering='average',
@@ -31,20 +18,21 @@ viewer.add_image(
     blending='translucent',
     opacity=0.5,
     depiction='plane',
-    plane=plane_parameters_z)
+    plane={
+        'position': (32, 32, 32),
+        'normal': (1, 0, 0),
+        'thickness': 10,
+    })
 
-viewer.add_image(
-    blobs,
-    name='blue plane',
-    rendering='average',
-    colormap='bop blue',
-    blending='additive',
-    opacity=0.5,
-    depiction='plane',
-    plane=plane_parameters_y)
+# create annotator
+annotator = n3d.annotators.SurfaceAnnotator(
+    viewer=viewer, image_layer=image_layer, enabled=True
+)
 
-spline_annotator = QtSurfaceAnnotatorWidget(viewer=viewer)
-viewer.window.add_dock_widget(spline_annotator)
-
-viewer.camera.angles = (60, 60, 60)
+# run napari
+viewer.layers.selection = [image_layer]
+viewer.axes.visible = True
+viewer.axes.labels = False
+viewer.camera.angles = (-15, 25, -30)
+viewer.camera.zoom *= 0.5
 napari.run()
