+++ src/napari_threedee/__init__.py
@@ -1,3 +1,7 @@
+from . import data_models
+from . import annotators
+from . import manipulators
+
 try:
     from ._version import version as __version__
 except ImportError:
