+++ src/napari_threedee/annotators/__init__.py
@@ -1,3 +1,4 @@
 from .points import PointAnnotator
 from .paths import PathAnnotator
 from .spheres import SphereAnnotator
+from .surfaces import SurfaceAnnotator
