+++ src/napari_threedee/annotators/surfaces/__init__.py
@@ -1,0 +1,1 @@
+from .annotator import SurfaceAnnotator
